+++ dgisim/src/card/cards_set.py
@@ -109,6 +109,7 @@ _DEFAULT_CARDS: list[type[Card]] = [
     ParametricTransformer,
     ## Support Card / Location ##
     KnightsOfFavoniusLibrary,
+    LiyueHarborWharf,
     Vanarana,
 ]