+++ dgisim/tests/test_card/test_liyue_harbor_wharf.py
@@ -1,0 +1,36 @@
+import unittest
+
+from .common_imports import *
+
+
+class TestLiyueHarborWharf(unittest.TestCase):
+    def test_behaviour(self):
+        base_state = ONE_ACTION_TEMPLATE
+        base_state = PublicAddCardEffect(Pid.P1, LiyueHarborWharf).execute(base_state)
+
+        game_state = base_state
+        game_state = step_action(game_state, Pid.P1, CardAction(
+            card=LiyueHarborWharf,
+            instruction=DiceOnlyInstruction(dices=ActualDices({Element.DENDRO: 2})),
+        ))
+
+        # test 2 more card is drawn for the first round
+        game_state = step_until_phase(game_state, game_state.get_mode().end_phase)
+        old_deck_cards = game_state.get_player1().get_deck_cards()
+        old_hand_cards = game_state.get_player1().get_hand_cards()
+        game_state = step_until_signal(game_state, TriggeringSignal.ROUND_END)
+        new_deck_cards = game_state.get_player1().get_deck_cards()
+        new_hand_cards = game_state.get_player1().get_hand_cards()
+        self.assertEqual(new_deck_cards + new_hand_cards, old_deck_cards + old_hand_cards)
+        self.assertEqual(new_hand_cards.num_cards(), old_hand_cards.num_cards() + 2)
+
+        # test 2 more card is drawn for the second round, and support disappears
+        game_state = step_until_phase(game_state, game_state.get_mode().end_phase)
+        old_deck_cards = game_state.get_player1().get_deck_cards()
+        old_hand_cards = game_state.get_player1().get_hand_cards()
+        game_state = step_until_signal(game_state, TriggeringSignal.ROUND_END)
+        new_deck_cards = game_state.get_player1().get_deck_cards()
+        new_hand_cards = game_state.get_player1().get_hand_cards()
+        self.assertNotIn(LiyueHarborWharfSupport, game_state.get_player1().get_supports())
+        self.assertEqual(new_deck_cards + new_hand_cards, old_deck_cards + old_hand_cards)
+        self.assertEqual(new_hand_cards.num_cards(), old_hand_cards.num_cards() + 2)